requirements.txt
speechrecognition
transformers
gtts
pyaudio
requests
bs4
google-auth
google-auth-oauthlib
google-api-python-client
assistant
